+++ test/use_case/view_favorites/ViewFavoritesInteractorTest.java
@@ -1,13 +1,12 @@
-package UseCase.ViewFavorites;
+package use_case.view_favorites;
 
-import Entities.User;
+import entities.User;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.mockito.Mock;
 import org.mockito.MockitoAnnotations;
 
 import java.time.LocalDateTime;
-import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
